requirements.txt
python-telegram-bot==13.15
telethon==1.28.5
python-dotenv==1.0.0
schedule==1.2.0
aiohttp==3.8.5
pydantic==2.1.1